+++ tests/fixtures/sql.py
@@ -134,3 +134,25 @@ import_operation = {
     "targetId": "pgsql-20191115184040913900000001",
     "targetProject": "chaosiqdemos"
 }
+
+
+databases = [
+  {
+    "kind": "sql#database",
+    "charset": "UTF8",
+    "collation": "en_US.UTF8",
+    "etag": "a797327e6265f5966f4995432791ab6cfdb09f47557d9339c0cef99a1ff3a011",
+    "name": "postgres",
+    "instance": "pgsql-20191115184040913900000001",
+    "project": "chaosiqdemos"
+  },
+  {
+    "kind": "sql#database",
+    "charset": "UTF8",
+    "collation": "en_US.UTF8",
+    "etag": "b905ba0bc656c7b4621d7733ea72f0b325f6e68c1f57ea1bd393aa3112ac9b1b",
+    "name": "chaos",
+    "instance": "pgsql-20191115184040913900000001",
+    "project": "chaosiqdemos"
+  }
+]